+++ docs/DeveloperGuide.md
@@ -939,7 +939,7 @@ There are two ways to run tests.
 
 This project has one type of test:
 Unit tests targeting the lowest level methods/classes.
-e.g. `seedu.duke.userinterface.command.notebook.AddNotebookTest`
+e.g. `seedu.duke.userinterface.command.notebook.AddNotebookModeTest`
 
 ## Appendix A: Project Scope
