+++ src/main/java/seedu/duke/exceptions/InvalidSelectCommandException.java
@@ -15,7 +15,7 @@ public class InvalidSelectCommandException extends ZeroNoteException {
     @Override
     public void printErrorMessage() {
         System.out.println("Invalid select command: " + argument);
-        System.out.println("Format for selecting: select /nNOTEBOOK /sSECTION /pNUMBER");
+        System.out.println("Format for selecting: select /nNOTEBOOK /sSECTION /pPAGE");
         System.out.println("You must select a notebook before selecting a section.");
         System.out.println("You must select a section before selecting a page.");
     }
